fix: adapt hardwareId,nativeId to be non required fields
The new Xesar API has a change that breaks its compatibility with xesar-connect when querying identification media.
Since Xesar 3.2 now supports smartphones as media, they are also included in the identification media query.
But they don't have a hardwareId and nativeId. This leads to a crash in xesar connect when the query response is deserialized because these fields are currently required fields.